requirements.txt
appdirs==1.4.4
asgiref==3.2.7
attrs==19.3.0
Babel==2.8.0
black==19.10b0
certifi==2020.4.5.1
chardet==3.0.4
click==7.1.2
coreapi==2.3.3
coreschema==0.0.4
dj-database-url==0.5.0
Django==3.0.7
django-cors-headers==3.3.0
django-filter==2.3.0
django-heroku==0.3.1
django-mysql==3.5.0
django-phonenumber-field==4.0.0
django-rest-swagger==2.2.0
djangorestframework==3.11.0
drf-yasg==1.17.1
gunicorn==20.0.4
heroku==0.1.4
idna==2.9
inflection==0.4.0
itypes==1.2.0
Jinja2==2.11.2
jsonschema==3.2.0
MarkupSafe==1.1.1
nose==1.3.7
openapi-codec==1.3.2
packaging==20.4
pathspec==0.8.0
phonenumbers==8.12.5
psycopg2==2.8.5
pyparsing==2.4.7
pyrsistent==0.16.0
python-dateutil==1.5
pytz==2020.1
PyYAML==5.3.1
regex==2020.5.14
requests==2.23.0
ruamel.yaml==0.16.10
ruamel.yaml.clib==0.2.0
simplejson==3.17.0
six==1.15.0
sqlparse==0.3.1
swagger-spec-validator==2.6.0
toml==0.10.1
typed-ast==1.4.1
uritemplate==3.0.1
urllib3==1.25.9
whitenoise==5.1.0
